Make dir optional (default "."), add return try to builtins, move resolvePath
- dir parameter is now last with default "." so read("foo.txt") works
- Builtins (read, write, readImage, fetch, fetchJSON) now use return try
  to consistently return failure Results instead of throwing
- resolvePath moved to its own file (stdlib/lib/resolvePath.ts) for
  discoverability
- Empty string dir still rejected with helpful error message
- Test call sites simplified: read("foo.txt") instead of read(".", "foo.txt")
