+++ packages/agency-lang/tests/agency-js/stdlib/std-fs-edit/agent.agency
@@ -2,7 +2,7 @@ import { edit } from "std::fs"
 
 node writeFixture(filename: string, content: string) {
   handle {
-    write(".", filename, content)
+    write(filename, content)
   } with (data) {
     return approve()
   }
@@ -11,7 +11,7 @@ node writeFixture(filename: string, content: string) {
 
 node readBack(filename: string) {
   handle {
-    const result = read(".", filename)
+    const result = read(filename)
   } with (data) {
     return approve()
   }
@@ -20,7 +20,7 @@ node readBack(filename: string) {
 
 node runEdit(filename: string, oldText: string, newText: string, replaceAll: boolean) {
   handle {
-    const result = edit(".", filename, oldText, newText, replaceAll)
+    const result = edit(filename, oldText, newText, replaceAll)
   } with (data) {
     return approve()
   }
