+++ packages/agency-lang/tests/agency-js/stdlib/std-fs-multiedit/agent.agency
@@ -1,12 +1,12 @@
 import { multiedit } from "std::fs"
 
 node writeFixture(filename: string, content: string) {
-  write(".", filename, content) with approve
+  write(filename, content) with approve
   return true
 }
 
 node readBack(filename: string) {
-  const result = read(".", filename) with approve
+  const result = read(filename) with approve
   return result
 }
 
@@ -17,6 +17,6 @@ type Edit = {
 }
 
 node runMultiedit(filename: string, edits: Edit[]) {
-  const result = multiedit(".", filename, edits) with approve
+  const result = multiedit(filename, edits) with approve
   return result
 }
